requirements.txt
# QA AI Automation — Dependencies

# AI (at least one required)
google-generativeai>=0.8.0   # Free tier — default provider
anthropic>=0.39.0            # Paid — used if GEMINI_API_KEY is empty
Pillow>=10.0.0               # Image handling for Gemini vision

# HTTP
requests>=2.31.0
httpx>=0.27.0

# Web server (webhook listener)
fastapi>=0.115.0
uvicorn>=0.32.0

# Configuration
python-dotenv>=1.0.0

# HTML parsing (Confluence content)
beautifulsoup4>=4.12.0

# Utilities
pydantic>=2.9.0

# Format learning (optional — for --format-file feature)
pandas>=2.2.0
openpyxl>=3.1.0
